gtp: fix protocol type handling in gtp_rx
The Rx has for a long time made the assumtion that the external
transport type is the same as the encapsulated data. It also has
silently assumed that the external was always IPv4.

Make the IPv4 only support for the encapsulated packet explicit by
hard coding it into the iptunnel_pull_header call.

check_src_ms also needs to know the encapsulated type. Moving it
after the iptunnel_pull_header means that it will now see the
skb->protocol value that has been updated to the inner protocol
type (and not as before the out tunnel transport type).

Since iptunnel already checked and pulled the no longer needed UDP
and GTP transport headers, the may_pull check can be removed.